Extract watched times export/import logic into useWatchedTimesSync hook
